+++ lib/src/data/_all.dart
@@ -1,3 +1,4 @@
 export 'enums/_all.dart';
 export 'models/_all.dart';
 export 'repositories/_all.dart';
+export 'logger.dart';