+++ lib/src/data/enums/_all.dart
@@ -1,4 +1,3 @@
 export 'flugger_data_type.dart';
-export 'flugger_folder_naming_convention.dart';
 export 'flugger_model_type.dart';
 export 'flugger_structure.dart';