+++ lib/src/generators/response_model_flugger_generator.dart
@@ -5,5 +5,6 @@ class ResponseModelFluggerGenerator extends ModelFluggerGenerator {
   ResponseModelFluggerGenerator({
     required super.options,
     required super.modelOptions,
+    required super.logger,
   });
 }