+++ lib/src/options/_all.dart
@@ -1,5 +1,3 @@
 export 'enum_flugger_options.dart';
 export 'flugger_options.dart';
 export 'model_flugger_options.dart';
-export 'structure_flugger_options.dart';
-export 'swagger_flugger_options.dart';